release: multisig trigger uint128 fix, router upgrade tooling, Safenet brief
Squash-merge of `dev` into `main`. Three logical groups of changes.

Multisig inheritance trigger (correctness)
- `IMultisigLegacyContract.setActivationTrigger` parameter widened from
  uint256 to uint128 to match `MultisigLegacyStruct.LegacyExtraConfig`;
  `MultisigLegacyContractRouter` now casts explicitly with an overflow
  guard. Resolves the ABI mismatch that caused signed-but-unexecuted
  multisig Safe transactions to revert and go un-indexed by the subgraph.
- New `test/MultisigActivationTrigger.spec.ts` covers the trigger path,
  the uint128 boundary, and the activation-time arithmetic.
- Existing `Legacy`, `PremiumAutomation`, `PremiumRegistry`, and
  `TimeLockRouter` test suites hardened: cleaner signer model (no
  hardcoded impersonation), consistent `dev` ownership, and a fix to
  `EIP712LegacyVerifier.setRouterAddresses` wiring in PremiumAutomation
  (multisig router slot was previously pointing at the EOA router).
- New `MockSafeWalletWithGuard` to exercise Guard interactions in tests.

Deploy / upgrade tooling
- `scripts/deploy-multisig-router-impl.ts` — deterministic build + deploy
  of a new `MultisigLegacyContractRouter` implementation for upgrade via
  the proxy admin.
- `scripts/upgrade-multisig-legacy-router.ts` — proxy-admin-driven upgrade
  flow with a pre-flight dry-run and post-upgrade sanity reads.
- `scripts/delete-eoa-legacy.ts` — operator utility used during the
  Sepolia QA to unblock stuck EOA test accounts.
- `scripts/sync-ui.ts` updated to emit the refreshed address set to the
  frontend and admin panel without regenerating unrelated entries.

Docs & outreach
- `docs/SAFENET_INTEGRATION_BRIEF.md` — discussion draft for Safe
  Foundation, Safenet Validators, and Transaction Checkers covering our
  on-chain footprint, the TransactionGuard collision between our
  `SafeGuard` and Safenet's Guard, four resolution paths (guard chaining,
  drop-our-guard, Safenet-aware wrapper, ModuleGuard-only), audit link,
  and contact points (info@/security@ 10102.io).
- `docs/plans/solc-upgrade.md` — working plan for the forthcoming solc
  version bump.

Addresses & tooling
- `contract-addresses.json` refreshed with the upgraded
  `MultisigLegacyRouter` implementation and assorted Sepolia / mainnet
  corrections.
- Hardhat config and package bumps consistent with the new deploy scripts.

Made-with: Cursor

+++ contracts/interfaces/IMultisigLegacyContract.sol
@@ -21,7 +21,7 @@ interface IMultisigLegacy {
     string[] calldata nicknames_
   ) external returns (uint256 numberOfBeneficiaries);
 
-  function setActivationTrigger(address sender_, uint256 lackOfOutgoingTxRange_) external;
+  function setActivationTrigger(address sender_, uint128 lackOfOutgoingTxRange_) external;
 
   function activeLegacy(address guardAddress_) external returns (address[] memory newSigners, uint256 newThreshold);
